Check for large mismatch in input and output geneset size

+++ msigdb.R
@@ -27,6 +27,7 @@ msigdb_url_base <- "https://data.broadinstitute.org/gsea-msigdb/msigdb"
 
 gencode_gtf <- str_glue("gencode.v{gencode_release}.basic.annotation.gtf")
 msigdb_db <- str_glue("msigdb_v{msigdb_release}.{msigdb_species}.db")
+msigdb_log <- str_glue("msigdb_v{msigdb_release}.{msigdb_species}.gencode.v{gencode_release}.basic.annotation.log")
 
 if (species == "human") {
   msigdb_chip <- str_glue("Human_Ensembl_Gene_ID_MSigDB.v{msigdb_release}.{msigdb_species}.chip")
@@ -51,11 +52,13 @@ if(!file.exists(gencode_gtf_gz_destfile)) download.file(url = gencode_gtf_gz_url
 if(!file.exists(msigdb_db_destfile) && !file.exists(msigdb_db_zip_destfile)) download.file(url = msigdb_db_zip_url, destfile = msigdb_db_zip_destfile)
 if(!file.exists(msigdb_chip_destfile)) download.file(url = msigdb_chip_url, destfile = msigdb_chip_destfile)
 
-# unzip zipped file
-unzip(msigdb_db_zip_destfile, exdir = here("input"))
+if(file.exists(msigdb_db_zip_destfile)) {
+  # unzip zipped file
+  unzip(msigdb_db_zip_destfile, exdir = here("input"))
 
-# remove zip file
-file.remove(msigdb_db_zip_destfile)
+  # remove zip file
+  file.remove(msigdb_db_zip_destfile)
+}
 
 # process gencode gtf
 gtf <- rtracklayer::import(gzfile(gencode_gtf_gz_destfile))
@@ -64,14 +67,14 @@ gencode.genes <- gtf |> filter(type == "gene") |> as_tibble()
 
 stopifnot(nrow(gencode.genes) == nrow(distinct(gencode.genes, gene_id)))
 
-gencode.genes.autosomal <- gencode.genes |> filter(seqnames %in% paste0("chr", 1:22), str_ends(gene_id, "_PAR_Y", negate = TRUE))
+gencode.genes <- gencode.genes |> filter(str_ends(gene_id, "_PAR_Y", negate = TRUE)) # seqnames %in% paste0("chr", 1:22)
 
-gencode.genes.autosomal |>
+gencode.genes |>
   group_by(gene_type) |>
   summarize(number = n()) |>
   print(n = Inf)
 
-ensgenes <- gencode.genes.autosomal |>
+ensgenes <- gencode.genes |>
   select(gene_id, gene_name) |>
   separate(gene_id, into = c("ensgene", NA), sep = "\\.", remove = FALSE)
 
@@ -103,21 +106,37 @@ gene_set_details <- tbl(db, "gene_set_details") |> as_tibble()
 
 DBI::dbDisconnect(db)
 
+# size of genesets in input
+gene_set_size_input <- gene_set |>
+  inner_join(gene_set_gene_symbol, join_by(id == gene_set_id)) |> 
+  count(standard_name)
+
 # join tables
-d <- gene_set |>
+gene_set <- gene_set |>
   inner_join(gene_set_gene_symbol, join_by(id == gene_set_id)) |> 
   inner_join(gene_symbol, join_by(gene_symbol_id == id)) |> 
   inner_join(ensgenes, join_by(symbol == gene_name), relationship = "many-to-many") |>
   inner_join(gene_set_details, join_by(id == gene_set_id)) |>
-  select(standard_name, systematic_name, collection_name, description_brief, symbol, NCBI_id, gene_id, ensgene) |>
+  select(id, standard_name, systematic_name, collection_name, description_brief, symbol, NCBI_id, gene_id, ensgene) |>
   separate(collection_name, into = "collection_name_top", sep = ":", remove = FALSE, extra = "drop")
 
-d
+gene_set
+
+# size of genesets in output
+gene_set_size_output <- gene_set |>
+  count(standard_name)
+
+# size of genesets in input and output, check for large mismatch
+gene_set_size <- gene_set_size_input |> 
+  left_join(gene_set_size_output, by = "standard_name", suffix = c("_input", "_output")) |>
+  filter(abs(n_input - n_output) >= (n_input * 0.2)) |>
+  write_tsv(msigdb_log) |> 
+  print(n = Inf)
 
 # write GMT files to output dir
-for (cn in unique(d$collection_name)) {
+for (cn in unique(gene_set$collection_name)) {
   suffix <- str_replace_all(cn, "[[:punct:]]", "\\.")
-  d |> filter(collection_name == cn) |>
+  gene_set |> filter(collection_name == cn) |>
     select(standard_name, description_brief, gene_id) |>
     arrange(standard_name, gene_id) |>
     nest(data = gene_id) |>
@@ -126,9 +145,9 @@ for (cn in unique(d$collection_name)) {
     write_tsv(str_glue("output/msigdb_v{msigdb_release}.{msigdb_species}.{suffix}.gmt"), col_names = FALSE)
 }
 
-for (cnt in unique(d$collection_name_top)) {
+for (cnt in unique(gene_set$collection_name_top)) {
   suffix = cnt
-  d |> filter(collection_name_top == cnt) |>
+  gene_set |> filter(collection_name_top == cnt) |>
     select(standard_name, description_brief, gene_id) |>
     arrange(standard_name, gene_id) |>
     nest(data = gene_id) |>
